Security fixes

+++ src/main/java/com/exadelinternship/carpool/services/CarService.java
@@ -3,10 +3,13 @@ package com.exadelinternship.carpool.services;
 import com.exadelinternship.carpool.adapters.CarAdapter;
 import com.exadelinternship.carpool.dto.CarDTO;
 import com.exadelinternship.carpool.entity.Car;
+import com.exadelinternship.carpool.entity.impl.UserDetailsImpl;
 import com.exadelinternship.carpool.repository.CarRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,20 +23,25 @@ public class CarService {
 
     public List<CarDTO> getAllCars(){
         List<CarDTO> result=new ArrayList<>();
-        carRepository.findAll().forEach(car -> result.add(carAdapter.carToCarDto(car)));
+        long userId=((UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
+        carRepository.findAllByUser_Id(userId).forEach(car -> result.add(carAdapter.carToCarDto(car)));
         return result;
     }
 
-    public CarDTO getCarById(long id){
-        return carAdapter.carToCarDto(carRepository.getOne(id));
-    }
-
+    @Transactional
     public void saveCar(CarDTO carDTO){
-        Car car=carAdapter.carDtoToCar(carDTO);
+        long userId=((UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
+        Car car=carAdapter.carDtoToCar(carDTO,userId);
         carRepository.save(car);
     }
 
+    @Transactional
     public void deleteCarById(long id){
-        carRepository.deleteById(id);
+        long userId=((UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
+        Car car=carRepository.getOne(id);
+        if(car.getUser().getId()==userId) {
+            carRepository.deleteById(id);
+        }
+
     }
 }
